Add Brotli support (#857)
It turns out that the Wayback Machine is serving us some content with Brotli encoding, even if we don't say we support it, leading to us storing and trying to analyze comprssed data as if it was already decompressed. Requests will use Brotli as long as it is installed, so installing it is all we need to do here.