Fix response body reading
I messed this up when I got record offsets. Turns out you have to read the body *first*, otherwise the stream is consumed and the body is lost (you can probably seek back on the file, but I have no idea what havoc that would wreak on the ArchiveIterator instance).