Handle Amazon WAF challenge interstitials
Amazon CloudFront Web Application Firewall works by responding to the browser with a page that does some JS fetches and adds a cookie, then reloads the page. We whould usually have captured the second request for the same URL, so just treat this like a redirect when resolving a chain of redirects and we should wind up with the real page instead of the challenge page.