Make cchardet optional
Cchardet is multiple orders of magnitude faster than chardet (and other similar pure Python tools for this purpose), which is why we chose to use it, but it later went unmaintained for several years. This has left us in a situation where we can no longer upgrade Python versions in a lot of our tools, since cchardet only works up to Python 3.10.

This makes cchardet an optional extra and otherwise uses charset_normalizer (which is pure Python, so should always work). This is a huge slowdown, but it appears we don't actually fall back to it very often in actual usage, so it's not the end of the world.

This also upgrades the version of cchardet we use to a pre-release. About a year ago, the maintainer started working on it again and released this preview release that seems to work alright up to Python 3.12. But it doesn't support Python 3.13 and work on it seems to have dropped off. Not sure what the real state of affairs is, so it's still best to keep this optional.