Make cchardet optional (#866)
Relying on cchardet is preventing us from supporting Python 3.11+, and that is becoming more and more of a problem. (We used it because it was *very* fast, but it has had serious maintenance issues for a few years now.)

This makes cchardet an optional extra and otherwise uses charset_normalizer (which is pure Python, so should always work). It’s a huge slowdown, but it appears we don't actually fall back to it very often in actual usage, so it's not the end of the world.

This also upgrades the optional version of cchardet we use to a pre-release. About a year ago, the maintainer started working on it again and released this preview that seems to work alright up to Python 3.12. But it doesn't support Python 3.13 and work on it seems to have dropped off. Not sure what the real state of affairs is, so it's still best to keep this optional.