Drop docopt (#867)
Docopt was a neat tool, but it imposed some weird constraints and edge cases on argument parsing and ordering, and extracting and parsing the values from it was often overcomplicated. The built-in argparse module doesn't have those issues and, while its syntax isn't as nice and pretty, it does more for us and is more concise.