Hide progress bar for WARC import on non-TTY (#868)
The progress bar can be a useful indicator, but when we run this job batched on non-interactive sessions, it makes it hard to see the warnings and errors since they show up in the middle of progress output (on an actual TTY this is no big deal, the progress bar is only one line and warnings move it down instead of showing up in the middle).

We can get it back in cases where we might need it by passing `--progress`.