Handle unknown PDF parsing errors
We currently handle `PyPdfError` exceptions when parsing PDFs. However, I ran into a malformed PDF that totally broke our crawl handling today because it raises a `ValueError`. Even though this is handled better in newer versions of pypdf that we aren't on yet, it still seems like this kind of thing is always reasonably possible - PDF parsing is just a fiddly and error-prone process, especially if we happen to get some garbled data.