Normalize URLs during WARC import
It turns out we've been missing some records because redirect URLs (in the 3xx response's `Location` header) do not always exactly match the requested target URL as recorded in the WARC record. This solves the problem by normalizing URLs before trying to match them.

For example, `https://www.heat.gov/` recently stopped getting recorded because it started redirecting to `https://heat.gov`, which got recorded in the WARC as `https://heat.gov/`. No path is exactly equivalent to `/` (note this is not the case for other trailing slashes in paths!), but because we were looking for a URL without a path, we didn't find a record, even though the correct matching record was there, just with a `/` for its path.

While working on this particular case, I noted some similar ones arising from variations on this same issue, like having redundant ports (e.g. the `443` is redundant here bcause it is an `https` scheme: `https://whatever.com:443/`).